Conflicts:
	src/scripts/InstanceScripts/Raid_OnyxiasLair.cpp
	src/scripts/InstanceScripts/Raid_SerpentshrineCavern.cpp
	src/scripts/InstanceScripts/Raid_TheObsidianSanctum.cpp

+++ src/scripts/InstanceScripts/Instance_TheVioletHold.cpp
@@ -103,7 +103,7 @@ public:
     {
         m_numBarrel = 0;
 
-        for (int i = 0; i < TVH_END; ++i)
+        for (uint8 i = 0; i < TVH_END; ++i)
             m_phaseData[i] = State_NotStarted;
 
         this->RegisterScriptUpdateEvent();
